requirements.txt
PyYAML
Jinja2
requests==2.32.5
